prepare_spec: Improve the license and group comments handling
Comments appearing between Summary and License/Group should not be
re-arranged since they usually refer to the tags below them. This
improves the situation by examining the lines following the Summary
tag and if it's a comment (or a sequense of comments) they are printed
before the License and Group tags.

Fixes: #13
Signed-off-by: Markos Chandras <mchandras@suse.de>