Do not capitalize URL/Url tags in premable
The two tools spec-cleaner and format_spec_file were both formatting the Url:
tag, but in different style. format_spec_file used Url, spec-cleaner URL, which results
in the two tools fighting over who is right, which in turn greatly annoys users.

Since spec-cleaner is the 'user requested' version of spec-cleaner (we install format_spec_file
bv default), assume that users of that tool want it that way. Thus we ignore the Url tag here
completely and don't reformat it.