Keep 3rd party licenses intact
Spec files from 3rd party sources - like Fedora - may carry a
license that must not be removed.
When parsing a spec file we assume that any line following a 3rd
party copyright that is a comment or empty is part of a 3rd party
copyright. It will be left in place.
This will inevitably catch also normal non-copyright comments.
These may appear before the SUSE license and terms if these are
missing and have to be added at the end.

Signed-off-by: Egbert Eich <eich@suse.com>