Fix _split_final_comments hack
The hack introduced in 5d082d1f to add trailing comments to
descriptions as comments to the next section never bothered
to use the proper methods to add lines to sections.
Also, it never bothered to handle trailing empty lines in
the newly created 'comment' nor in the modified 'description'
section properly.
This lead to the script not being idempotent, ie. consecutive
passes could each produce a different result.

This fix addresses this problem.

Signed-off-by: Egbert Eich <eich@suse.com>